backend/requirements.txt
django
django-environ
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-jazzmin==2.4.8
django-ckeditor
django-dbbackup
django-q
django-celery-beat
django-celery-results
django-redis
django-rq
django-extensions
Pillow
psycopg2-binary
gunicorn
six
typing-extensions
flake8
black
isort
toml
wheel
twilio
redis
celery
PyYAML
watchdog
drf-yasg
jsonschema
sentry-sdk
razorpay
django-redis-cache